4단계/refact : carName을 vo로 변경
