+++ src/main/java/domain/CarName.java
@@ -1,0 +1,42 @@
+package domain;
+
+import java.util.Objects;
+
+public class CarName {
+    private final String name;
+
+    public CarName(String name) {
+        validateCarName(name);
+        this.name = name;
+    }
+
+    private void validateCarName(String name) {
+        if (name == null || name.isBlank())
+            throw new IllegalArgumentException("유효하지 않은 입력입니다.");
+        if (name.length() > 5) {
+            throw new InvalidLengthUserNameException("이름은 5자까지 가능합니다.");
+        }
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (!(o instanceof CarName carName)) return false;
+        return Objects.equals(name, carName.name);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(name);
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    static class InvalidLengthUserNameException extends IllegalArgumentException {
+        InvalidLengthUserNameException(String errorMessage) {
+            super(errorMessage);
+        }
+    }
+}
